Move menu config out of views.py
